코드 스타일 개선

+++ src/main/java/com/studiop/dormmanagementsystem/service/BuildingService.java
@@ -4,6 +4,7 @@ import com.studiop.dormmanagementsystem.entity.Building;
 import com.studiop.dormmanagementsystem.entity.dto.BuildingDto;
 import com.studiop.dormmanagementsystem.entity.enums.FridgeType;
 import com.studiop.dormmanagementsystem.repository.BuildingRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,11 +13,18 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class BuildingService {
 
     private final BuildingRepository buildingRepository;
 
+    public Building getById(Long id) {
+        return buildingRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("해당 ID의 건물 찾을 수 없습니다: " + id));
+    }
+
+    @Transactional
     public Building addBuilding(String name, int fridgeSlots, int freezerSlots, int integratedSlots, FridgeType type) {
         return buildingRepository.save(new Building(name, fridgeSlots, freezerSlots, integratedSlots, type));
     }
@@ -34,8 +42,7 @@ public class BuildingService {
 
     @Transactional
     public void editBuilding(Long id, String buildingName, int fridgeSlots, int freezerSlots, int integratedSlots, FridgeType type) {
-        Building building = buildingRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Building not found with id: " + id));
+        Building building = getById(id);
 
         building.changeName(buildingName);
         building.changeFridgeSlots(fridgeSlots);
@@ -46,6 +53,9 @@ public class BuildingService {
 
     @Transactional
     public void deleteBuilding(Long id) {
+        if (!buildingRepository.existsById(id)) {
+            throw new EntityNotFoundException("해당 ID의 건물을 찾을 수 없습니다: " + id);
+        }
         buildingRepository.deleteById(id);
     }
 }
